refactor(workflow): reorganize workflow utility imports and enhance structure
